+++ src/test/java/com/iota/iri/hash/ISSTests.java
@@ -1,5 +1,6 @@
 package com.iota.iri.hash;
 
+import com.iota.iri.utils.Converter;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -7,14 +8,14 @@ import static org.junit.Assert.*;
 /**
  * Created by paul on 2/13/17.
  */
-public class ISCTest {
+public class ISSTests {
     @Test
     public void encrypt() throws Exception {
         final String key, plain, cipher;
         key = "ABCDEFGHIJKLM9NOPQRSTUVWXYZ";
         plain = "9NOPQRSTUVWXYZABCDEFGHIJKLM";
         cipher = "ASRQYXWDCBJIHZOQSUWY9BDFHJL";
-        assert(ISC.encrypt(key).apply(plain).equals(cipher));
+        assert(Converter.trytes(ISS.encrypt(Converter.trits(key)).apply(Converter.trits(plain))).equals(cipher));
     }
 
     @Test
@@ -23,7 +24,7 @@ public class ISCTest {
         key = "ABCDEFGHIJKLM9NOPQRSTUVWXYZ";
         plain = "9NOPQRSTUVWXYZABCDEFGHIJKLM";
         cipher = "ASRQYXWDCBJIHZOQSUWY9BDFHJL";
-        assert(ISC.decrypt(key).apply(cipher).equals(plain));
+        assert(Converter.trytes(ISS.decrypt(Converter.trits(key)).apply(Converter.trits(cipher))).equals(plain));
     }
 
 }
